+++ packages/techdocs-cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @techdocs/cli
 
+## 1.8.15-next.0
+
+### Patch Changes
+
+- c964a3d: Import discovery from backend-defaults instead of backend-common
+- Updated dependencies
+  - @backstage/backend-defaults@0.3.3-next.0
+  - @backstage/plugin-techdocs-node@1.12.7-next.0
+  - @backstage/catalog-model@1.5.0
+  - @backstage/cli-common@0.1.14
+  - @backstage/config@1.2.0
+
 ## 1.8.12
 
 ### Patch Changes