+++ plugins/example-todo-list-backend/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @internal/plugin-todo-list-backend
 
+## 1.0.29-next.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.21-next.0
+  - @backstage/backend-common@0.23.2-next.0
+  - @backstage/plugin-auth-node@0.4.16-next.0
+  - @backstage/errors@1.2.4
+
 ## 1.0.28
 
 ### Patch Changes